상품-카테고리 레포작업

+++ src/main/java/com/spharosacademy/project/SSGBack/product/repository/CategoryLRepository.java
@@ -3,5 +3,8 @@ package com.spharosacademy.project.SSGBack.product.repository;
 import com.spharosacademy.project.SSGBack.product.domain.CategoryL;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryLRepository extends JpaRepository<CategoryL, Long> {
+import java.util.List;
+
+
+public interface CategoryLRepository extends JpaRepository<CategoryL, Integer> {
 }
